# frontend/docs-backend/openapi_QUYNH_v8.0.0.yaml
...
openapi: 3.1.0
info:
  title: Quynh Web API
  description: Cổng thông tin điện tử nội bộ Lữ đoàn Thông tin 21. Xem lịch sử thay
    đổi ở openapi.CHANGELOG.md.
  version: 8.0.0
paths:
...
  /units/:
    post:
      tags:
      - units
      summary: Create Unit
      operationId: create_unit_units__post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnitCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - units
      summary: List Units
      operationId: list_units_units__get
      security:
      - HTTPBearer: []
      parameters:
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 200
          title: Limit
      - name: kind
        in: query
        required: false
        schema:
          anyOf:
          - enum:
            - phong_ban
            - tieu_doan
            - dai_doi
            - tram
            - bch_lu_doan
            - cap_uy
            type: string
          - type: 'null'
          title: Kind
      - name: active
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Active
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UnitOut'
                title: Response List Units Units  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /units/{unit_id}:
    get:
      tags:
      - units
      summary: Get Unit
      operationId: get_unit_units__unit_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: unit_id
        in: path
        required: true
        schema:
          type: integer
          title: Unit Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - units
      summary: Update Unit
      operationId: update_unit_units__unit_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: unit_id
        in: path
        required: true
        schema:
          type: integer
          title: Unit Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnitCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - units
      summary: Delete Unit
      operationId: delete_unit_units__unit_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: unit_id
        in: path
        required: true
        schema:
          type: integer
          title: Unit Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /users/{user_id}/unit:
    patch:
      tags:
      - users
      summary: Set Unit
      operationId: set_unit_users__user_id__unit_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnitAssignUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /profile/me:
    get:
      tags:
      - profile
      summary: Get My Profile
      operationId: get_my_profile_profile_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
      security:
      - HTTPBearer: []
    put:
      tags:
      - profile
      summary: Update My Profile
      operationId: update_my_profile_profile_me_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /profile/permissions:
    get:
      tags:
      - profile
      summary: Get My Permissions
      description: Tra ve ma tran phan quyen chi tiet va Khoi/Nganh co quan cua quan
        nhan.
      operationId: get_my_permissions_profile_permissions_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPermissionsOut'
      security:
      - HTTPBearer: []
  /profile/change-password:
    post:
      tags:
      - profile
      summary: Change My Password
      operationId: change_my_password_profile_change_password_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChange'
        required: true
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /chats/{id}/messages/upload:
    post:
      tags:
      - chats
      summary: Send Message With File
      description: 'Gửi tin nhắn kèm tệp đính kèm vào cuộc trò chuyện, đẩy thời gian
        thực tới thành viên online.


        Định dạng cho phép: ảnh (.jpg/.jpeg/.png/.webp/.gif), tài liệu

        (.pdf/.doc(x)/.xls(x)/.ppt(x)), video (.mp4/.webm/.ogg/.mov/.m4v) và

        tệp nén (.zip/.rar/.7z/.tar/.gz/.bz2). Video và tệp nén dùng hạn mức

        `MAX_VIDEO_UPLOAD_MB`, còn lại `MAX_UPLOAD_MB`. Sai định dạng / quá dung

        lượng / tệp rỗng → 400; không phải thành viên hội thoại → 403.'
      operationId: send_message_with_file_chats__id__messages_upload_post
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_send_message_with_file_chats__id__messages_upload_post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/upload:
    post:
      tags:
      - upload
      summary: Upload File
      operationId: upload_file_api_upload_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_file_api_upload_post'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    Body_send_message_with_file_chats__id__messages_upload_post:
      properties:
        content:
          type: string
          title: Content
          description: Chú thích kèm tệp (tuỳ chọn)
          default: ''
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
          description: Ảnh / tài liệu / video đính kèm
      type: object
      required:
      - file
      title: Body_send_message_with_file_chats__id__messages_upload_post
...
    Body_upload_file_api_upload_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - file
      title: Body_upload_file_api_upload_post
...
    ChatMessageOut:
      properties:
        id:
          type: integer
          title: Id
        conversation_id:
          type: integer
          title: Conversation Id
        sender_id:
          type: integer
          title: Sender Id
        sender_name:
          type: string
          title: Sender Name
        sender_rank:
          anyOf:
          - type: string
          - type: 'null'
          title: Sender Rank
        sender_position:
          anyOf:
          - type: string
          - type: 'null'
          title: Sender Position
        content:
          type: string
          title: Content
        attachment_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Attachment Url
        attachment_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Attachment Name
        created_at:
          type: string
          format: date-time
          title: Created At
        is_me:
          type: boolean
          title: Is Me
          default: false
      type: object
      required:
      - id
      - conversation_id
      - sender_id
      - sender_name
      - content
      - created_at
      title: ChatMessageOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PasswordChange:
      properties:
        current_password:
          type: string
          title: Current Password
        new_password:
          type: string
          minLength: 1
          title: New Password
      type: object
      required:
      - current_password
      - new_password
      title: PasswordChange
...
    ProfileUpdate:
      properties:
        full_name:
          type: string
          maxLength: 100
          title: Full Name
      type: object
      required:
      - full_name
      title: ProfileUpdate
...
    UnitAssignUpdate:
      properties:
        unit_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Unit Id
      type: object
      title: UnitAssignUpdate
    UnitCreate:
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 2
          title: Name
        unit_kind:
          type: string
          enum:
          - phong_ban
          - tieu_doan
          - dai_doi
          - tram
          - bch_lu_doan
          - cap_uy
          title: Unit Kind
          default: phong_ban
        description:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Description
        is_active:
          type: boolean
          title: Is Active
          default: true
      type: object
      required:
      - name
      title: UnitCreate
    UnitOut:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        unit_kind:
          type: string
          enum:
          - phong_ban
          - tieu_doan
          - dai_doi
          - tram
          - bch_lu_doan
          - cap_uy
          title: Unit Kind
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        is_active:
          type: boolean
          title: Is Active
        created_at:
          type: string
          format: date-time
          title: Created At
        user_count:
          type: integer
          title: User Count
      type: object
      required:
      - id
      - name
      - unit_kind
      - description
      - is_active
      - created_at
      - user_count
      title: UnitOut
    UploadOut:
      properties:
        status:
          type: string
          const: success
          title: Status
          default: success
        filename:
          type: string
          title: Filename
        url:
          type: string
          title: Url
      type: object
      required:
      - filename
      - url
      title: UploadOut
      description: Ket qua tra ve cho POST /api/upload (upload file dung chung).
...
    UserOut:
      properties:
        id:
          type: integer
          title: Id
        username:
          type: string
          title: Username
        full_name:
          type: string
          title: Full Name
        role:
          type: integer
          title: Role
        role_label:
          type: string
          title: Role Label
        rank:
          anyOf:
          - type: string
          - type: 'null'
          title: Rank
        position:
          anyOf:
          - type: string
          - type: 'null'
          title: Position
        is_active:
          type: boolean
          title: Is Active
        clearance:
          type: boolean
          title: Clearance
        unit_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Unit Id
        unit_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Unit Name
        directive_channel_access:
          type: boolean
          title: Directive Channel Access
        command_channel_access:
          type: boolean
          title: Command Channel Access
        is_system:
          type: boolean
          title: Is System
        must_change_password:
          type: boolean
          title: Must Change Password
      type: object
      required:
      - id
      - username
      - full_name
      - role
      - role_label
      - is_active
      - clearance
      - unit_id
      - directive_channel_access
      - command_channel_access
      - is_system
      - must_change_password
      title: UserOut
...
    UserPermissionsOut:
      properties:
        user_id:
          type: integer
          title: User Id
        username:
          type: string
          title: Username
        full_name:
          type: string
          title: Full Name
        role:
          type: integer
          title: Role
        role_label:
          type: string
          title: Role Label
        branch:
          type: string
          title: Branch
        branch_label:
          type: string
          title: Branch Label
        unit_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Unit Id
        unit_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Unit Name
        permissions:
          additionalProperties:
            type: boolean
          type: object
          title: Permissions
      type: object
      required:
      - user_id
      - username
      - full_name
      - role
      - role_label
      - branch
      - branch_label
      - permissions
      title: UserPermissionsOut
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer